Revert reorganization of DWCTheme components.
