+++ docs/src/components/DWCTheme/ColorPalette/ColorPalette.css
