+++ docs/src/theme/MDXComponents.js
@@ -20,18 +20,15 @@ import AccordionSummary from '@mui/material/AccordionSummary';
 import AccordionDetails from '@mui/material/AccordionDetails';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import GLOBALS from '../../siteConfig';
+import RadiusBox from '@site/src/components/DWCTheme/RadiusBox/RadiusBox';
+import ShadowBox from '@site/src/components/DWCTheme/ShadowBox/ShadowBox';
+import SizingBox from '@site/src/components/DWCTheme/SizingBox/SizingBox';
+import SpacingBox from '@site/src/components/DWCTheme/SpacingBox/SpacingBox';
+import SurfaceBox from '@site/src/components/DWCTheme/SurfaceBox/SurfaceBox';
+import TransitionBox from '@site/src/components/DWCTheme/TransitionBox/TransitionBox';
 import GalleryCard from '@site/src/components/GalleryCard/GalleryCard';
 import GalleryGrid from '@site/src/components/GalleryGrid/GalleryGrid';
-
-import {
-  ColorPalette,
-  RadiusBox,
-  ShadowBox,
-  SizingBox,
-  SpacingBox,
-  SurfaceBox,
-  TransitionBox
-} from '@site/src/components/DWCTheme/DWCTheme.js';
+import ColorPalette from '@site/src/components/DWCTheme/ColorPalette/ColorPalette';
 
 
 export default {
